+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -9,6 +9,8 @@ import java.net.Socket;
 
 import org.apache.catalina.HandlerMapping;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.request.Http11RequestReader;
+import org.apache.coyote.http11.response.Http11ResponseWriter;
 import org.apache.coyote.response.Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
