+++ tomcat/src/main/java/org/apache/coyote/http11/request/Http11RequestLineReader.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import static java.util.stream.Collectors.*;
